[ALPS03648512] FPSGO: coverity
[Detail]
Unintentional integer overflow

[Solution]
To avoid overflow, set dur to type unsigned long long

MTK-Commit-Id: 3feba19833ca02464b310fe3ecd994ffe00897af

CR-Id: ALPS03648512
Feature: FPSGO

+++ drivers/misc/mediatek/performance/fpsgo_v2/fbt/src/fbt_cpu.c
@@ -1630,7 +1630,7 @@ static unsigned long long fbt_est_loading(int cur_ts, atomic_t last_ts, unsigned
 		return 0;
 
 	if (cur_ts > atomic_read(&last_ts)) {
-		int dur = cur_ts - atomic_read(&last_ts);
+		unsigned long long dur = cur_ts - atomic_read(&last_ts);
 
 		return dur * obv;
 	}
